// src/web3/dto/sc-item.dto.ts
// itemId: '100',
//   nft: '0x5FbDB2315678afecb367f032d93F642f64180aa3',
//   tokenId: '100',
//   price: '55000000000000000000',
//   listingPrice: '55000000000000000000',
//   seller: '0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266',
//   isSold: false

export class ScItemDto {
  itemId: string;
  nft: string;
  tokenId: string;
  price: string;
  listingPrice: string;
  seller: string;
  isSold: boolean;
}